Fix linguistics issues and unclear statements in requirements
The user case study part is still not clear (is this submitted in a
written document?)

Why is this a user case study and not just a explanation and
justification of their work? Perhaps make explicit that they need to
explain their application by describing what a user would want to use it
for and the steps that the user would
go through.